+++ plugins/woocommerce-admin/client/task-list/task-list.js
@@ -53,6 +53,21 @@ export const TaskList = ( {
 		possiblyTrackCompletedTasks();
 	}, [ query ] );
 
+	const visibleTasks = tasks.filter(
+		( task ) => task.visible && ! dismissedTasks.includes( task.key )
+	);
+
+	const completedTaskKeys = visibleTasks
+		.filter( ( task ) => task.completed )
+		.map( ( task ) => task.key );
+
+	const incompleteTasks = tasks.filter(
+		( task ) =>
+			task.visible &&
+			! task.completed &&
+			! dismissedTasks.includes( task.key )
+	);
+
 	const possiblyCompleteTaskList = () => {
 		const taskListVariableName = `woocommerce_${ name }_complete`;
 		const taskListToComplete = isComplete
@@ -64,8 +79,8 @@ export const TaskList = ( {
 		}
 
 		if (
-			( ! getIncompleteTasks().length && ! isComplete ) ||
-			( getIncompleteTasks().length && isComplete )
+			( ! incompleteTasks.length && ! isComplete ) ||
+			( incompleteTasks.length && isComplete )
 		) {
 			updateOptions( {
 				...taskListToComplete,
@@ -73,26 +88,11 @@ export const TaskList = ( {
 		}
 	};
 
-	const getCompletedTaskKeys = () => {
-		return getVisibleTasks()
-			.filter( ( task ) => task.completed )
-			.map( ( task ) => task.key );
-	};
-
-	const getIncompleteTasks = () => {
-		return tasks.filter(
-			( task ) =>
-				task.visible &&
-				! task.completed &&
-				! dismissedTasks.includes( task.key )
-		);
-	};
-
 	const getTrackedIncompletedTasks = (
 		partialCompletedTasks,
 		allTrackedTask
 	) => {
-		return getVisibleTasks()
+		return visibleTasks
 			.filter(
 				( task ) =>
 					allTrackedTask.includes( task.key ) &&
@@ -102,7 +102,6 @@ export const TaskList = ( {
 	};
 
 	const possiblyTrackCompletedTasks = () => {
-		const completedTaskKeys = getCompletedTaskKeys();
 		const trackedCompletedTasks = getTrackedCompletedTasks(
 			completedTaskKeys,
 			totalTrackedCompletedTasks
@@ -160,12 +159,6 @@ export const TaskList = ( {
 		} );
 	};
 
-	const getVisibleTasks = () => {
-		return tasks.filter(
-			( task ) => task.visible && ! dismissedTasks.includes( task.key )
-		);
-	};
-
 	const recordTaskListView = () => {
 		if ( query.task ) {
 			return;
@@ -173,8 +166,6 @@ export const TaskList = ( {
 
 		const isCoreTaskList = name === 'task_list';
 		const taskListName = isCoreTaskList ? 'tasklist' : 'extended_tasklist';
-
-		const visibleTasks = getVisibleTasks();
 
 		recordEvent( `${ taskListName }_view`, {
 			number_tasks: visibleTasks.length,
@@ -196,8 +187,8 @@ export const TaskList = ( {
 
 		recordEvent( `${ taskListName }_completed`, {
 			action,
-			completed_task_count: getCompletedTaskKeys().length,
-			incomplete_task_count: getIncompleteTasks().length,
+			completed_task_count: completedTaskKeys.length,
+			incomplete_task_count: incompleteTasks.length,
 		} );
 		updateOptions( {
 			...updateOptionsParams,
@@ -223,7 +214,7 @@ export const TaskList = ( {
 		);
 	};
 
-	const listTasks = getVisibleTasks().map( ( task ) => {
+	const listTasks = visibleTasks.map( ( task ) => {
 		if ( ! task.onClick ) {
 			task.onClick = ( e ) => {
 				if ( e.target.nodeName === 'A' ) {
@@ -277,7 +268,7 @@ export const TaskList = ( {
 					<CardHeader size="medium">
 						<div className="wooocommerce-task-card__header">
 							<Text variant="title.small">{ listTitle }</Text>
-							<Badge count={ getIncompleteTasks().length } />
+							<Badge count={ incompleteTasks.length } />
 						</div>
 						{ renderMenu() }
 					</CardHeader>
@@ -294,6 +285,9 @@ export const TaskList = ( {
 									onDismiss={ () => dismissTask( task ) }
 									time={ task.time }
 									level={ task.level }
+									action={ task.onClick }
+									actionLabel={ task.action }
+									additionalInfo={ task.additionalInfo }
 								/>
 							) ) }
 						</ListComp>